pyproject.toml
[project]
name = "personal-llm"
version = "0.1.0"
description = "Personalized LLM with RAG on WhatsApp messages using Ollama"
authors = [
    {name = "Phil Nash",email = "noemail@example.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "ollama (>=0.6.1,<0.7.0)",
    "langchain (>=1.2.4,<2.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "chromadb (>=1.4.1,<2.0.0)",
    "sentence-transformers (>=5.2.0,<6.0.0)",
    "langchain-text-splitters (>=1.1.0,<2.0.0)",
    "langchain-ollama (>=1.0.1,<2.0.0)",
    "langchain-huggingface (>=1.2.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
